tools/simulator/pyproject.toml
[project]
name = "p1-simulator"
version = "0.1.0"
description = "Synthetic user behavior generator for P1 e-commerce platform"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "pyarrow>=18.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
]
kafka = [
    "aiokafka>=0.11.0",
]

[project.scripts]
p1-simulator = "p1_simulator.run_simulator:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
